Move fixtures to conftest.py, resolve F401 in tests.
