+++ tests/test_gctrace.py
@@ -4,7 +4,6 @@ from tests.schemas import gctrace_chromtab
 from tests.utils import (
     datagram_from_input,
     standard_datagram_test,
-    datadir,
     compare_result_dicts,
 )
